packages/celery-redis/pyproject.toml
[project]
name = "celery-redis"
version = "0.1.0"
description = "프로젝트에서 사용하는 공통 Redis 관련 메소드"
requires-python = ">=3.13"
dependencies = [
    "redis>=5.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.build]
module-root = "src"

[build-system]
requires = ["uv_build>=0.5.18"]
build-backend = "uv_build"
